move SocketHandler to data package
