+++ service_location/src/main/kotlin/org/thechance/service_location/di/AppModule.kt
@@ -1,7 +1,7 @@
 package org.thechance.service_location.di
 
 import org.koin.dsl.module
-import org.thechance.service_location.util.SocketHandler
+import org.thechance.service_location.data.SocketHandler
 
 val appModule = module {
     single<SocketHandler> { SocketHandler() }
